+++ src/main/java/insane96mcp/survivalreimagined/data/generator/SRRecipeProvider.java
@@ -367,7 +367,7 @@ public class SRRecipeProvider extends RecipeProvider implements IConditionBuilde
                         400
                 )
                 .experience(1f)
-                .doubleOutputChance(0.2f)
+                .outputIncrease(0.2f)
                 .unlockedBy("has_scrap_block", has(Durium.SCRAP_BLOCK.item().get()))
                 .save(writer, SurvivalReimagined.RESOURCE_PREFIX + "durium_ingot_from_soul_blasting");
 
@@ -465,7 +465,7 @@ public class SRRecipeProvider extends RecipeProvider implements IConditionBuilde
                         400
                 )
                 .experience(2f)
-                .doubleOutputChance(0.2f)
+                .outputIncrease(0.2f)
                 .unlockedBy("has_hellish_coal", has(CoalFire.HELLISH_COAL.get()))
                 .save(writer, SurvivalReimagined.RESOURCE_PREFIX + "soul_steel_ingot_from_soul_blasting");