+++ src/main/java/insane96mcp/survivalreimagined/module/mining/multiblockfurnaces/data/MultiItemSmeltingRecipeBuilder.java
@@ -31,7 +31,7 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
     protected String group;
     final NonNullList<Ingredient> ingredients;
     float experience = 0f;
-    float doubleOutputChance = 0f;
+    float outputIncrease = 0f;
     private final Item result;
     protected final int cookingTime;
     @Nullable
@@ -73,8 +73,8 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
         return this;
     }
 
-    public MultiItemSmeltingRecipeBuilder doubleOutputChance(float doubleOutputChance) {
-        this.doubleOutputChance = doubleOutputChance;
+    public MultiItemSmeltingRecipeBuilder outputIncrease(float outputIncrease) {
+        this.outputIncrease = outputIncrease;
         return this;
     }
 
@@ -96,7 +96,7 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
     public void save(Consumer<FinishedRecipe> pFinishedRecipeConsumer, ResourceLocation pRecipeId) {
         this.ensureValid(pRecipeId);
         this.advancement.parent(ROOT_RECIPE_ADVANCEMENT).addCriterion("has_the_recipe", RecipeUnlockedTrigger.unlocked(pRecipeId)).rewards(AdvancementRewards.Builder.recipe(pRecipeId)).requirements(RequirementsStrategy.OR);
-        pFinishedRecipeConsumer.accept(new MultiItemSmeltingRecipeBuilder.Result(pRecipeId, this.group == null ? "" : this.group, this.bookCategory, this.ingredients, this.result, this.experience, this.doubleOutputChance, this.cookingTime, this.recycle, this.advancement, pRecipeId.withPrefix("recipes/" + this.category.getFolderName() + "/"), this.serializer));
+        pFinishedRecipeConsumer.accept(new MultiItemSmeltingRecipeBuilder.Result(pRecipeId, this.group == null ? "" : this.group, this.bookCategory, this.ingredients, this.result, this.experience, this.outputIncrease, this.cookingTime, this.recycle, this.advancement, pRecipeId.withPrefix("recipes/" + this.category.getFolderName() + "/"), this.serializer));
     }
 
     /**
@@ -115,7 +115,7 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
         private final NonNullList<Ingredient> ingredients;
         private final Item result;
         private final float experience;
-        private final float doubleOutputChance;
+        private final float outputIncrease;
         private final int cookingTime;
         @Nullable
         private final AbstractMultiItemSmeltingRecipe.Recycle recycle;
@@ -123,14 +123,14 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
         private final ResourceLocation advancementId;
         private final RecipeSerializer<? extends AbstractMultiItemSmeltingRecipe> serializer;
 
-        public Result(ResourceLocation pId, String pGroup, SRBookCategory pCategory, NonNullList<Ingredient> pIngredients, Item pResult, float pExperience, float doubleOutputChance, int pCookingTime, @Nullable AbstractMultiItemSmeltingRecipe.Recycle recycle, Advancement.Builder pAdvancement, ResourceLocation pAdvancementId, RecipeSerializer<? extends AbstractMultiItemSmeltingRecipe> pSerializer) {
+        public Result(ResourceLocation pId, String pGroup, SRBookCategory pCategory, NonNullList<Ingredient> pIngredients, Item pResult, float pExperience, float outputIncrease, int pCookingTime, @Nullable AbstractMultiItemSmeltingRecipe.Recycle recycle, Advancement.Builder pAdvancement, ResourceLocation pAdvancementId, RecipeSerializer<? extends AbstractMultiItemSmeltingRecipe> pSerializer) {
             this.id = pId;
             this.group = pGroup;
             this.category = pCategory;
             this.ingredients = pIngredients;
             this.result = pResult;
             this.experience = pExperience;
-            this.doubleOutputChance = doubleOutputChance;
+            this.outputIncrease = outputIncrease;
             this.cookingTime = pCookingTime;
             this.recycle = recycle;
             this.advancement = pAdvancement;
@@ -153,8 +153,8 @@ public class MultiItemSmeltingRecipeBuilder implements RecipeBuilder {
             pJson.addProperty("result", ForgeRegistries.ITEMS.getKey(this.result).toString());
             if (experience > 0f)
                 pJson.addProperty("experience", this.experience);
-            if (doubleOutputChance > 0f)
-                pJson.addProperty("double_output_chance", this.doubleOutputChance);
+            if (outputIncrease > 0f)
+                pJson.addProperty("output_increase", this.outputIncrease);
             if (this.recycle != null)
                 pJson.add("recycle", this.recycle.toJson());
             pJson.addProperty("cookingtime", this.cookingTime);